catalog: extract catalog_ingest.py — import-time model ingestion (#38 step 3)
Move the 12 ingestion symbols (OpenRouter-snapshot + provider-manifest parsing,
embedding seed, endpoint builders + their private constants) into
catalog_ingest.py. Verified clean: they're pure PRODUCERS (build model/endpoint
dicts from catalog_data + pricing), never read the live MODELS/MODEL_ENDPOINTS
registries, so no import cycle. catalog.py imports them and keeps the import-time
merge/mutation side-effects (it owns the registries). Facade re-exports.

catalog.py: 3374 -> 1590 lines (more than halved across pricing+data+ingest).
MODELS still builds 270/701; full suite 1175 green, mypy clean.
